Added M1A1 Thompson Weapon

+++ Assets/Scripts/Actors/Health.cs
@@ -10,14 +10,14 @@ using UnityEngine;
 
 public class Health : MonoBehaviour //Parent to PlayerInteractions and Trap, as both objects have health that can be decremented.
 {
-    [SerializeField] protected int maxHealth, currentHealth; //So child classes can access it.
+    [SerializeField] protected float maxHealth, currentHealth; //So child classes can access it.
 
     void Start() //sets current health to max health at start.
     {
         currentHealth = maxHealth;
     }
 
-    public virtual void TakeDamage(int damageInflicted) //Called by other scripts, inflicts damage and checks if the object needs to die.
+    public virtual void TakeDamage(float damageInflicted) //Called by other scripts, inflicts damage and checks if the object needs to die.
     {
         currentHealth -= damageInflicted;
